requirements.txt
# Run with: pip install -r requirements.txt
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.2.5
asttokens==2.0.5
attrs==21.4.0
av==11.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.0
Django==3.0.4
docopt==0.6.2
docutils==0.16
entrypoints==0.4
executing==0.8.2
filelock==3.0.12
Fiona==1.8.21
fonttools==4.49.0
fsspec==2024.2.0
GDAL==3.4.3
geopandas==0.11.1
idna==3.3
importlib-resources==5.4.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
Js2Py==0.71
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
kiwisolver==1.4.5
MarkupSafe==2.0.1
matplotlib==3.7.5
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.3.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.1
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pillow==10.2.0
pipwin==0.5.2
platformdirs==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
pure-eval==0.2.2
pycparser==2.21
Pygments==2.6.1
pyjsparser==2.7.1
pyparsing==3.0.7
pypiwin32==223
PyPrind==2.11.3
pyproj==3.4.0
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
Send2Trash==1.8.0
Shapely==1.8.4
six==1.14.0
soupsieve==2.3.2.post1
split-folders==0.5.1
sqlparse==0.3.1
stack-data==0.1.4
sympy==1.12
terminado==0.13.1
testpath==0.5.0
tomli==2.0.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.1
traitlets==5.1.1
tweepy==4.5.0
typing_extensions==4.9.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.8
virtualenv==20.0.13
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0